Persistence update - the server is now able to save the state of the game at the end of every turn. When the game crashes (check not yet implemented) the existence of a saved game is useful: if all the players restart the program and 1. Log with the same name 2. The previous admin creates a game with the same number of players
The admin is prompted to load the game. If the answer is positive the game is reloaded from disk, otherwise a new game is created.

The saving procedure ensures that the last saved state is always the most updated, and that when the game ends an already started turn is not considered.

The chat history is not subject to save, as it can be easily recreated by the players, who already know what they were talking about.

* in order to implement this functionality, the main model classes were subject to a small refactor, and methods converting the game objects containing important data were added in GameObjectConverter.