Adjust final turns logic and bug fixes relative to loaded games.
Being "player X" the player who has completed the shelf before everyone else,
continue the game until the player at the right of the player X has completed his turn, or if the player X is the first player.

Corrected the bug relative to the missed end game trigger after a game load with a full shelf.